parse this_class, super_class, interfaces_count

+++ src/main/java/xyz/muscaestar/zero2hp/bytecode/classfile/item/constantpool/struct/CONSTANT_Class_info.java
@@ -26,6 +26,6 @@ public class CONSTANT_Class_info extends CpInfo {
 
     @Override
     public String meta() {
-        return "[2字节]name_index: #" + name_index;
+        return "[2字节]name_index: #" + (int) name_index;
     }
 }
